Unhappy class TX deleted

+++ src/vlc_rm/transmitter.py
@@ -48,12 +48,12 @@ class Transmitter:
         if mlambert <= 0:
             raise ValueError("Lambert number must be greater than zero.")        
 
-        self._wavelengths = np.array(wavelengths,  dtype=np.float32)
+        self._wavelengths = np.array(wavelengths, dtype=np.float32)
         if self._wavelengths.size != Kt.NO_LEDS:
             raise ValueError(
                 "Dimension of wavelengths array must be equal to the number of LEDs.")
 
-        self._fwhm = np.array(fwhm)
+        self._fwhm = np.array(fwhm, dtype=np.float32)
         if self._fwhm.size != Kt.NO_LEDS:
             raise ValueError(
                 "Dimension of FWHM array must be equal to the number of LEDs.")
